+++ framework/CHANGELOG.md
@@ -37,6 +37,7 @@ Yii Framework 2 Change Log
 - Enh #8670: Added support for saving extra fields in session table for `yii\web\DbSession` (klimov-paul)
 - Chg #6354: `ErrorHandler::logException()` will now log the whole exception object instead of only its string representation (cebe)
 - Chg #8556: Extracted `yii\web\User::getAuthManager()` method (samdark)
+- Bug #8901: Fixed `yii\web\AssetManager` can not get correct published path or URL when only partial published (hightman)
 
 
 2.0.4 May 10, 2015